refactor(recall): extract RecallInternalDeps into a sibling module
Pure mechanical extraction, no behavior change. Moves the
`RecallInternalDeps` interface (the deps contract for the recall-internal
coordinator, issue #1526 seam 18) verbatim from recall-internal.ts into a
new sibling `recall-internal-deps.ts`, and drops the imports that were
only consumed by that interface (now unused at the call site).

Why: recall-internal.ts sits at its frozen ratchet ceiling (5674 lines),
which blocks recall work landing in #2183 — any net growth fails
`check-ratchets` and the baseline cannot be raised. The ratchet tool's
own prescribed remedy is "extract the addition into a sibling module."
This gives recall-internal.ts ~359 lines of headroom so the connector-
label threading in #2183 can land without touching the baseline. New
sibling modules are uncapped.

No renames, no member reordering, no logic changes. The interface block
is moved byte-for-byte; recall-internal.ts imports it back.

Part of #2183
